performance on Intel E5-2697A 32 cores
without patch
100G shmem_init:      25.32 secs
     shmem_finalize:   3.26 secs
200G shmem_init:     106.29 secs
     shmem_finalize    9.55 secs

with patch
100G shmem_init:       1.63 secs
     shmem_finalize:   0.44 secs
200G shmem_init:       3.01 secs
     shmem_finalize    0.85 secs
`