@mellanox-github 2019-10-07
Test FAILed.
See http://hpc-master.lab.mtl.com:8080/job/hpc-ucx-pr/12526/ for details (Mellanox internal link).