@mellanox-github 2019-10-07
<details><summary>Mellanox CI: <b>FAILED</b> on 25 of 29 workers (click for details)</summary>

**Note:** the logs will be deleted after 14-Oct-2019

Agent/Stage | Status
-- | --
[_main](https://gist.githubusercontent.com/mellanox-github/93a49edc80c4f1bcd824405604d7ca9c/raw/d297b40c108fd19415ffa7184a974e5450a24ffc/_main) | :x: FAILURE
[hpc-arm-cavium-jenkins_W0](https://gist.githubusercontent.com/mellanox-github/93a49edc80c4f1bcd824405604d7ca9c/raw/935a152674b02a282b7aa159438b3fa3f1ec1d43/hpc-arm-cavium-jenkins_W0) | :x: FAILURE
[hpc-arm-cavium-jenkins_W1](https://gist.githubusercontent.com/mellanox-github/93a49edc80c4f1bcd824405604d7ca9c/raw/abf932cb7755bcb4a8983bced183f01cf26aad7b/hpc-arm-cavium-jenkins_W1) | :x: FAILURE
[hpc-arm-cavium-jenkins_W2](https://gist.githubusercontent.com/mellanox-github/93a49edc80c4f1bcd824405604d7ca9c/raw/b660e35eaee61b8cb0bd423305805c651891560d/hpc-arm-cavium-jenkins_W2) | :x: FAILURE
[hpc-arm-cavium-jenkins_W3](https://gist.githubusercontent.com/mellanox-github/93a49edc80c4f1bcd824405604d7ca9c/raw/0429cc8a5cdeed33df89f4180b565d92db8a3b45/hpc-arm-cavium-jenkins_W3) | :x: FAILURE
[hpc-arm-hwi-jenkins_W0](https://gist.githubusercontent.com/mellanox-github/93a49edc80c4f1bcd824405604d7ca9c/raw/543ae8ee190443feae959800d4b292777e0f5d4a/hpc-arm-hwi-jenkins_W0) | :x: FAILURE
[hpc-arm-hwi-jenkins_W1](https://gist.githubusercontent.com/mellanox-github/93a49edc80c4f1bcd824405604d7ca9c/raw/32d7dc76df94213be62c08263217e23a2a42b728/hpc-arm-hwi-jenkins_W1) | :x: FAILURE
[hpc-arm-hwi-jenkins_W2](https://gist.githubusercontent.com/mellanox-github/93a49edc80c4f1bcd824405604d7ca9c/raw/30bde4d67178646fe1792e3306e03bcac6d53df7/hpc-arm-hwi-jenkins_W2) | :x: FAILURE
[hpc-arm-hwi-jenkins_W3](https://gist.githubusercontent.com/mellanox-github/93a49edc80c4f1bcd824405604d7ca9c/raw/9cf9a04983f57b0689d3731ad13cf78a31350d08/hpc-arm-hwi-jenkins_W3) | :x: FAILURE
[hpc-test-node-gpu_W0](https://gist.githubusercontent.com/mellanox-github/93a49edc80c4f1bcd824405604d7ca9c/raw/b7ec6bfcbb94c3be7f2dc50fb467cbe15027e07f/hpc-test-node-gpu_W0) | :x: FAILURE
[hpc-test-node-gpu_W1](https://gist.githubusercontent.com/mellanox-github/93a49edc80c4f1bcd824405604d7ca9c/raw/328b2c7fb3baaedbf7d9b477e897e07f9c7c6e4d/hpc-test-node-gpu_W1) | :x: FAILURE
[hpc-test-node-gpu_W2](https://gist.githubusercontent.com/mellanox-github/93a49edc80c4f1bcd824405604d7ca9c/raw/e0305ace9f1b60158573d6898d67f89f13a6a4f8/hpc-test-node-gpu_W2) | :x: FAILURE
[hpc-test-node-gpu_W3](https://gist.githubusercontent.com/mellanox-github/93a49edc80c4f1bcd824405604d7ca9c/raw/97e83564912804cf059882dd5cdbe7badd9c3a77/hpc-test-node-gpu_W3) | :x: FAILURE
[hpc-test-node-legacy_W0](https://gist.githubusercontent.com/mellanox-github/93a49edc80c4f1bcd824405604d7ca9c/raw/b979b60892381ef008b7f5ff7302ec1a53eb0d77/hpc-test-node-legacy_W0) | :x: FAILURE
[hpc-test-node-legacy_W1](https://gist.githubusercontent.com/mellanox-github/93a49edc80c4f1bcd824405604d7ca9c/raw/9ff2181b06edfc7748e1dc1a584b029b0867413b/hpc-test-node-legacy_W1) | :x: FAILURE
[hpc-test-node-legacy_W2](https://gist.githubusercontent.com/mellanox-github/93a49edc80c4f1bcd824405604d7ca9c/raw/a7a4f136f94664843e68222a621c404dc168a9a6/hpc-test-node-legacy_W2) | :x: FAILURE
[hpc-test-node-legacy_W3](https://gist.githubusercontent.com/mellanox-github/93a49edc80c4f1bcd824405604d7ca9c/raw/43b62dfb8a9e7304bf5b602c558cb8cb0cfc68f1/hpc-test-node-legacy_W3) | :x: FAILURE
[hpc-test-node-new_W0](https://gist.githubusercontent.com/mellanox-github/93a49edc80c4f1bcd824405604d7ca9c/raw/410a719c2841615f2ac0f1bdf2d7319ba7ca29e9/hpc-test-node-new_W0) | :x: FAILURE
[hpc-test-node-new_W1](https://gist.githubusercontent.com/mellanox-github/93a49edc80c4f1bcd824405604d7ca9c/raw/d810c53de9bc56777b3fae44468579b84e90a156/hpc-test-node-new_W1) | :x: FAILURE
[hpc-test-node-new_W2](https://gist.githubusercontent.com/mellanox-github/93a49edc80c4f1bcd824405604d7ca9c/raw/ba2e5d6e8bde8e38437dc294f7f897ba3d9e14a9/hpc-test-node-new_W2) | :x: FAILURE
[hpc-test-node-new_W3](https://gist.githubusercontent.com/mellanox-github/93a49edc80c4f1bcd824405604d7ca9c/raw/8768c21d52e29f81789324b8edcc49917229f6d2/hpc-test-node-new_W3) | :x: FAILURE
[r-vmb-ppc-jenkins_W0](https://gist.githubusercontent.com/mellanox-github/93a49edc80c4f1bcd824405604d7ca9c/raw/f9dad89b783adc8f38defc4c4af3181da6d10fcf/r-vmb-ppc-jenkins_W0) | :x: FAILURE
[r-vmb-ppc-jenkins_W1](https://gist.githubusercontent.com/mellanox-github/93a49edc80c4f1bcd824405604d7ca9c/raw/419e9548fa784151c4e653589b155fb9dd105ddc/r-vmb-ppc-jenkins_W1) | :x: FAILURE
[r-vmb-ppc-jenkins_W2](https://gist.githubusercontent.com/mellanox-github/93a49edc80c4f1bcd824405604d7ca9c/raw/3e58919ef68fe4323216231740ca74ed47ccfc27/r-vmb-ppc-jenkins_W2) | :x: FAILURE
[r-vmb-ppc-jenkins_W3](https://gist.githubusercontent.com/mellanox-github/93a49edc80c4f1bcd824405604d7ca9c/raw/6379bcdfcc3b0919d19c6cacf848c32b298fe0ef/r-vmb-ppc-jenkins_W3) | :x: FAILURE
[hpc-test-althca_W0](https://gist.githubusercontent.com/mellanox-github/93a49edc80c4f1bcd824405604d7ca9c/raw/98516df946cdc9c5e253f55c61d667aac2e9c6bc/hpc-test-althca_W0) | :heavy_check_mark: SUCCESS
[hpc-test-althca_W1](https://gist.githubusercontent.com/mellanox-github/93a49edc80c4f1bcd824405604d7ca9c/raw/68fdf8882e343fe52e3205d4f1b3dc48048299e9/hpc-test-althca_W1) | :heavy_check_mark: SUCCESS
[hpc-test-althca_W2](https://gist.githubusercontent.com/mellanox-github/93a49edc80c4f1bcd824405604d7ca9c/raw/d87794b82dfb3bf5b281f22c93460b58a237b60b/hpc-test-althca_W2) | :heavy_check_mark: SUCCESS
[hpc-test-althca_W3](https://gist.githubusercontent.com/mellanox-github/93a49edc80c4f1bcd824405604d7ca9c/raw/df84d36e87bfc01f0ece3460b14a4d7044aed24b/hpc-test-althca_W3) | :heavy_check_mark: SUCCESS